+++ packages/cms/config/contentWidgets.js
@@ -7,11 +7,16 @@
 const contentWidgets = {
   'agenda': {},
   'accordeon': {},
-  'arguments': {
+  'arguments-block': {
     adminOnly: true
   },
   'arguments-form': {
-    adminOnly: true
+    adminOnly: true,
+    deprecated: true
+  },
+  'arguments': {
+    adminOnly: true,
+    deprecated: true
   },
   'section': {
     addLabel: 'Columns',