+++ packages/cms/lib/modules/admin-widgets/index.js
@@ -8,16 +8,17 @@ module.exports = {
   extend: 'openstad-widgets',
   label: 'Admin panel',
   addFields: [
-  //  styleSchema.definition('containerStyles', 'Styles for the container')
+    //  styleSchema.definition('containerStyles', 'Styles for the container')
   ],
   construct: function(self, options) {
     self.apos.app.get('/admin', function(req, res) {
-        // Place any data you need to access in your template here:
-    //    req.data = {};
+      // Place any data you need to access in your template here:
+      //    req.data = {};
 
-        // self.sendPage is what will actually return the page and render the template 'profile.html' in your views folder.
-        // You can change 'profile' to the name of your template minus '.html' - e.g. 'page.html' would just be 'page'
-        return self.sendPage(req, '/page', {});
+      // self.sendPage is what will actually return the page and render the template 'profile.html' in your views folder.
+      // You can change 'profile' to the name of your template minus '.html' - e.g. 'page.html' would just be 'page'
+      let openstadReactAdminCdn = self.apos.settings.getOption(req, 'siteConfig').openstadReactAdminCdn;
+      return self.sendPage(req, '/page', {openstadReactAdminCdn});
     });
   }
 };