+++ packages/cms/lib/modules/counter-widgets/index.js
@@ -2,6 +2,7 @@
  * Widget displays a counter with dynamic count of votes, users voted and ideas submitted
  * Often used to display "120 ideas submitted" "3000 users voted"
  */
+const _ = require('lodash');
 const styleSchema = require('../../../config/styleSchema.js').default;
 
 module.exports = {
@@ -47,11 +48,13 @@ module.exports = {
 						'staticCount'
 					]
         },
-/*         {
-          label: 'Arguments count',
-          value: 'argumentsCount',
+        {
+          label: 'Argument count',
+          value: 'argumentCount',
+          showFields: [
+            'ideaId'
+          ]
         },
-        */
       ]
     },
     {
@@ -80,6 +83,12 @@ module.exports = {
       ],
       required: false
     },
+    {
+      name: 'id',
+      type: 'string',
+      label: 'idea ID - leave empty to fetch total arguments',
+      required: false,
+    },
     styleSchema.definition('containerStyles', 'Styles for the container')
   ],
   construct: function(self, options) {
@@ -89,8 +98,17 @@ module.exports = {
         widgets.forEach((widget) => {
 
           const siteConfig = req.data.global.siteConfig;
-          if (widget.counterType == 'voteCount') {
+          if (widget.counterType === 'voteCount') {
             widget.isCountPublic = siteConfig && siteConfig.votes && siteConfig.votes.isViewable ? siteConfig.votes.isViewable : false;
+          }
+          else if (widget.counterType === 'argumentCount') {
+            if (!widget.id) {
+              widget.argsCount = req.data.ideas ? _.reduce(req.data.ideas, (count, idea) => { return count + (idea.argCount ? idea.argCount : 0)}, 0) : false;
+            } else {
+              const idea = req.data.ideas ? req.data.ideas.filter(function (idea) { return idea.id == widget.id}) : [];
+              widget.argsCount = idea ? _.reduce(idea, (count, idea) => { return count + (idea.argCount ? idea.argCount : 0)}, 0) : false;
+            }
+            widget.isCountPublic = true;
           } else {
             widget.isCountPublic = true;
           }
@@ -103,9 +121,16 @@ module.exports = {
           widget.cssHelperClassesString = widget.cssHelperClasses ? widget.cssHelperClasses.join(' ') : '';
 
           widget.siteId = req.data.global.siteId;
+          
+          widget.parsedUrl = widget.url;
+          if (widget.parsedUrl) {
+            let match = widget.parsedUrl.match(/^http/);
+            if (!match) {
+              widget.parsedUrl = req.data.siteUrl + widget.parsedUrl;
+            }
+          }
+
         });
-
-
 
         return superLoad(req, widgets, function (err) {
             if (err) {
@@ -153,6 +178,10 @@ module.exports = {
           count = widget.staticCount;
           break;
 
+        case 'argumentCount':
+          count = widget.argsCount;
+          break;
+
         default:
           count = 0;
 
@@ -163,6 +192,7 @@ module.exports = {
 
       var result = superOutput(widget, options);
       return result;
+
     };
 
     const superPushAssets = self.pushAssets;