+++ packages/cms/lib/modules/ideas-on-map-widgets/lib/create-config.js
@@ -1,6 +1,7 @@
-const sortingOptions  = require('../../../../config/sorting.js').ideasOnMapOptions;
+const sortingOptions = require('../../../../config/sorting.js').ideasOnMapOptions;
+const ideaForm = require('./idea-form');
 
-module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos) {
+module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, imageProxy, apos) {
 
   let contentConfig = {
     ignoreReactionsForIdeaIds: widget.ignoreReactionsForIdeaIds,
@@ -19,15 +20,7 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
   // image settings; todo: deze moeten syncen naar de api en dan moet de voorwaardelijkheid omgedraaid
   let allowMultipleImages = typeof widget.imageAllowMultipleImages != 'undefined' ? widget.imageAllowMultipleImages : ( ( data.global.siteConfig && data.global.siteConfig.ideas && typeof data.global.siteConfig.ideas.allowMultipleImages != 'undefined' ) ? data.global.siteConfig.ideas.allowMultipleImages : false );
   let placeholderImageSrc = typeof widget.imagePlaceholderImageSrc != 'undefined' ? apos.attachments.url(widget.imagePlaceholderImageSrc) : ( ( data.global.siteConfig && data.global.siteConfig.ideas && typeof data.global.siteConfig.ideas.placeholderImageSrc != 'undefined' ) ? data.global.siteConfig.ideas.placeholderImageSrc : undefined );
-
-  // formfields
-  let formFields = [ ...widget.formFields ];
-  formFields.forEach((formField) => {
-    if ( formField.inputType ==  "image-upload" ) {
-      formField.allowMultiple = allowMultipleImages; // todo: ik dnek dat deze niet meer nodig is
-    }
-  });
-
+  
   let themeTypes;
   try {
     themeTypes = data.global.themes || [];
@@ -107,8 +100,8 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
 
     image: {
       server: {
-				process: '/image',
-				fetch: '/image',
+				process: imageProxy,
+				fetch: imageProxy,
         srcExtension: '/:/rs=w:[[width]],h:[[height]];cp=w:[[width]],h:[[height]]',
       },
       aspectRatio: widget.imageAspectRatio || '16x9',
@@ -128,7 +121,7 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
 			descriptionMinLength: ( data.global.siteConfig && data.global.siteConfig.ideas && data.global.siteConfig.ideas.descriptionMinLength ) || 30,
 			descriptionMaxLength: ( data.global.siteConfig && data.global.siteConfig.ideas && data.global.siteConfig.ideas.descriptionMaxLength ) || 200,
 			allowMultipleImages,
-      fields: formFields,
+      fields: ideaForm.getWidgetFormFields(widget),
       shareChannelsSelection: widget.showShareButtons ? widget.shareChannelsSelection : [],
       metaDataTemplate: widget.metaDataTemplate,
 		},
@@ -137,14 +130,14 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
 
     argument: {
       isActive: widget.showReactions,
-      isClosed: data.global.siteConfig && data.global.siteConfig.arguments && typeof data.global.siteConfig.arguments.isClosed != 'undefined' ? data.global.siteConfig.arguments.isClosed : false,
-      closedText: data.global.siteConfig && data.global.siteConfig.arguments && typeof data.global.siteConfig.arguments.closedText != 'undefined' ? data.global.siteConfig.arguments.closedText : true,
       title: widget.reactionsTitle,
       formIntro: widget.reactionsFormIntro,
       placeholder: widget.reactionsPlaceholder,
 			descriptionMinLength: ( data.global.siteConfig && data.global.siteConfig.arguments && data.global.siteConfig.arguments.descriptionMinLength ) || 30,
 			descriptionMaxLength: ( data.global.siteConfig && data.global.siteConfig.arguments && data.global.siteConfig.arguments.descriptionMaxLength ) || 100,
-      closeReactionsForIdeaIds: widget.closeReactionsForIdeaIds,
+      isClosed: typeof widget.reactionsClosed != 'undefined' ? !!widget.reactionsClosed : (data.global.siteConfig && data.global.siteConfig.arguments && typeof data.global.siteConfig.arguments.isClosed != 'undefined' ? data.global.siteConfig.arguments.isClosed : false),
+      closedText: typeof widget.reactionsClosedText != 'undefined' ? widget.reactionsClosedText : (data.global.siteConfig && data.global.siteConfig.arguments && typeof data.global.siteConfig.arguments.closedText != 'undefined' ? data.global.siteConfig.arguments.closedText : true),
+      closeReactionsForIdeaIds: widget.reactionsClosed === '' && widget.closeReactionsForIdeaIds || '',
 		},
 
     map: {
@@ -167,7 +160,10 @@ module.exports = function createConfig(widget, data, jwt, apiUrl, loginUrl, apos
       isActive: data.global.siteConfig.votes.isActive,
       isActiveFrom: data.global.siteConfig.votes.isActiveFrom,
       isActiveTo: data.global.siteConfig.votes.isActiveTo,
+      requiredUserRole: data.global.siteConfig.votes.requiredUserRole,
+      voteType: data.global.siteConfig.votes.voteType,
       voteValues: data.global.siteConfig.votes.voteValues,
+
     },
 
   }