+++ packages/cms/lib/modules/resource-representation-widgets/index.js
@@ -35,7 +35,6 @@ module.exports = {
           const globalData = req.data.global;
           const siteConfig = req.data.global.siteConfig;
 
-
           widgets.forEach((widget) => {
               // render string with variables. Add active recource
               if (widget.containerStyles) {
@@ -49,6 +48,8 @@ module.exports = {
               widget.mapCenterLat = globalData.mapCenterLat;
               widget.mapCenterLng = globalData.mapCenterLng;
               widget.mapPolygons = globalData.mapPolygons;
+
+              widget.countdownPeriod = siteConfig.ideas.automaticallyUpdateStatus && siteConfig.ideas.automaticallyUpdateStatus.afterXDays || 0;
 
               widget.siteConfig = {
                   minimumYesVotes: (siteConfig && siteConfig.ideas && siteConfig.ideas.minimumYesVotes),
@@ -78,6 +79,9 @@ module.exports = {
           const openStadMap =  widget.siteConfig && widget.siteConfig.openStadMap ? widget.siteConfig.openStadMap : {};
           const markerStyle = widget.siteConfig && widget.siteConfig.openStadMap && widget.siteConfig.openStadMap.markerStyle ? widget.siteConfig.openStadMap.markerStyle : null;
           const idea = widget.activeResource;
+
+          let daysOld = parseInt( ( Date.now() - new Date(idea.startDate).getTime() ) / ( 24 * 60 * 60 * 1000 ) );
+          idea.countdown = widget.countdownPeriod - daysOld;
 
           //map expects array
           const ideas = widget.activeResource ? [widget.activeResource] : [];